rename metrics test module
